[Perf] Variable shrink-wrapping based on live range analysis
Summary:
We normally output all of a function's variables at the top of a function. This
commit uses the live range information to compute the latest point at which a
variable can be declared. This can avoid the initialization cost in cases where
the paths where the variable is used are not excercised.

Test Plan:
fast_tests, slow_tests, flib/unit tests.

Reviewed By: qigao
Reviewers: myang, qigao
CC: ps, mwilliams, qigao
Differential Revision: 239213