[Fix] Casting to array, string and object was broken
Summary:
First, the various to* functions weren't all handling null correctly. In
addition, if we had an explicit cast in the code, but type inference deduced
that the type was correct, we omitted the cast. This is wrong, because when we
infer an array, it actually means array or null; and a cast of null to array
results in an empty array.

Test Plan:
fast_tests slow_tests

Reviewed By: stephentu
Reviewers: qigao, myang, stephentu
CC: ps, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 246993