[Perf] Switch statements with all string literals for cases are now optimized
Summary: A jump table much like the one used for method invocations is used to
reduce the number of times equals() is called for the switch operand.
A new method hashForStringSwitch() is added for Variants, Arrays, and Objects in
order to support this. Additionally, the jump table allows for traversing in the
order of case statements, which is necessary for Objects.

Task ID: #541854

Blame Rev:

Reviewers: mwilliams, qigao, myang

CC:

Test Plan: fast_tests, slow_tests

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 248851