[Fix] ObjectPropertyExpression::lval was broken
Summary:
An ObjectPropertyExpression whose m_obj was not an lvalue, could return a
reference to a variant which had already been destroyed.

The problem was that it created a variant to hold the evaluated object, then
called o_lval on that. But if the local variant was the only reference to the
evaluated object, it would be destroyed (taking the result of o_lval with it)
before lval returned.

The solution is to copy the value to a temporary location so it can persist
until the function returns (at which point the value will either be consumed and
discarded, or bound to a variant).

Test Plan:
fast_tests (with new test case added), and TestCodeRun-ObjectProperties

Reviewed By: qigao
Reviewers: qigao, myang
CC: ps, mwilliams, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 248971