[Fix] Fix sigma build failure
Summary:
The sigma build failed in the final linking step because the symbol
HPHP::SystemLib::AllocStdClassObject is missing. This symbol is defined
in the new file systemlib.cpp. We need to include it into libcore.a.

Test Plan:
build libcore and verify the symbol is defined.

Reviewed By: qigao
Reviewers: qigao, andrewparoski
CC: hphp-diffs@lists, stein, astrehl, ps, mwilliams, myang, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 250192