[Perf] Faster code for checkClassExists
Summary:
checkClassExists was an out-of-line function. It took a pointer to the flag
indicating the presnce of the class, to allow fast checks for the class, and a
bool indicating whether or not to call __autoload, rather than having to call
function_exists, and a boolean to indicate whether or not to throw.

I've partially inlined the function. The fast check is now inlined, and if it
fails, it calls out to either a throw, or no-throw version of the function. Ive
dropped the autoload flag, since its rarely needed (due to spl_autoload). This
means that the usual path (where the class does exist) is now much shorter, and
doesnt involve any branches. When the class doesnt exist, the call only has two
parameters, so involves fewer instructions. Overall the code size goes up a
little, but the extra code is in the "cold" part of the function.

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: qigao, myang
CC: ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 255182