[Perf] Return ObjectData* from coo_ functions
Summary:
The coo_* functions were returning Object. This was good for the co_*
functions, and neutral for the i_dummy case, which accounted for a large
proportion of the uses. It was awkward when using it as a replacement for
NEWOBJ. Investigation showed that the co_ methods are no longer needed, so I
removed them, and changed coo_X* to return an ObjectData*, making it a drop-in
replacement for NEWOBJ(X).

The result is a decent code size reduction from removing the co_* functions,
and slightly smaller/faster code for "new", and "clone".

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: myang, qigao
CC: cbueno, ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 258322