[Fix] The local expression altered bit was being incorrectly computed for function calls
Summary:
For something like:

  g($d, f($a, $b, $c));

when markLocalExprAltered() was call on f, it would check
for interference with $a, $b, $c, and $d because they
were ahead of f on the access chain. However, we really
only want to check for intereference with $d in this case.
We were already keeping track of the end of an expression
on the access chain, so I added functionality to keep track
of the beginning of where to start checking for interference.

Test Plan:
slow_tests

Reviewed By: mwilliams
Reviewers: mwilliams
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 258852